Possibility to add video

+++ Main/views.py
@@ -20,7 +20,7 @@ def card_detail(request, card_id):
                       'card': card,
                       'tags': [tag.name for tag in card.tags.all()],
                       'all_tags': [tag.name for tag in Tag.objects.all()],
-                      'image_parts': [part.to_dict() for part in card.imagepart_set.all()]
+                      'image_parts': [part.to_dict() for part in card.imagepart_set.all()],
                   })
 
 
@@ -58,7 +58,8 @@ def save_card(request):
         title = data.get('title', 'base')
         description = data.get('description', None)
         tags = data.get('tags', None)
-        image_data = data.get('image', None)
+        content_data = data.get('content', None)
+        card_type = data.get('card_type', None)  # може бути "image" або "video"
         image_parts = data.get('image_parts', None)
 
         logging.info(f'card_id: {card_id}; title: {title}; description: {description}; tags: {tags} image_parts: {image_parts}')
@@ -70,16 +71,21 @@ def save_card(request):
 
         card.title = title
         card.description = description
+        card.type = 1 if card_type == "image" else 2
 
-        if image_data:
-            format, imgstr = image_data.split(';base64,')
+        if content_data:
+            format, content_str = content_data.split(';base64,')
             ext = format.split('/')[-1]
 
-            data = ContentFile(base64.b64decode(imgstr), name='temp.' + ext)
-            card.image.save(str(uuid.uuid4()), data, save=True)
+            data = ContentFile(base64.b64decode(content_str), name='temp.' + ext)
+
+            if card_type == "image":
+                card.image.save(str(uuid.uuid4()), data, save=True)
+            elif card_type == "video":
+                card.video.save(str(uuid.uuid4()), data, save=True)
         else:
             if not card_id:
-                return JsonResponse({"message": "Image is required."}, status=400)
+                return JsonResponse({"message": "Content (image or video) is required."}, status=400)
 
         card.save()
 
@@ -88,41 +94,42 @@ def save_card(request):
         card.tags.clear()
         card.add_tags_and_parents(tags)
 
-        # Image parts tags validation
-        for image_part_dct in image_parts:
-            image_part_tags = image_part_dct['tags']
-            image_part_tags = [tag.strip() for tag in image_part_tags]
-            logging.info(f'image_part_tags: {image_part_tags}')
+        if card_type == "image":
+            # Image parts tags validation
+            for image_part_dct in image_parts:
+                image_part_tags = image_part_dct['tags']
+                image_part_tags = [tag.strip() for tag in image_part_tags]
+                logging.info(f'image_part_tags: {image_part_tags}')
 
-            for tag_got in image_part_tags:
-                tag = Tag.objects.filter(name=tag_got).first()
-                logging.info(f'tag: {tag_got} {tag}')
-                if not tag:
-                    return JsonResponse({"message": f"Tag '{tag_got}' does not exist."}, status=400)
+                for tag_got in image_part_tags:
+                    tag = Tag.objects.filter(name=tag_got).first()
+                    logging.info(f'tag: {tag_got} {tag}')
+                    if not tag:
+                        return JsonResponse({"message": f"Tag '{tag_got}' does not exist."}, status=400)
 
-        # Delete old image parts
-        card.imagepart_set.all().delete()
+            # Delete old image parts
+            card.imagepart_set.all().delete()
 
-        added_parts = 0
-        for image_part_dct in image_parts:
-            image_part = ImagePart()
-            image_part.card = card
-            image_part.start_x = image_part_dct['start_x']
-            image_part.start_y = image_part_dct['start_y']
-            image_part.width = image_part_dct['width']
-            image_part.height = image_part_dct['height']
+            added_parts = 0
+            for image_part_dct in image_parts:
+                image_part = ImagePart()
+                image_part.card = card
+                image_part.start_x = image_part_dct['start_x']
+                image_part.start_y = image_part_dct['start_y']
+                image_part.width = image_part_dct['width']
+                image_part.height = image_part_dct['height']
 
-            image_part.save()
+                image_part.save()
 
-            image_part_tags = image_part_dct['tags']
-            image_part_tags = [tag.strip() for tag in image_part_tags]
-            image_part_tags = Tag.objects.filter(name__in=image_part_tags)
-            image_part.tags.add(*image_part_tags)
-            added_parts += 1
+                image_part_tags = image_part_dct['tags']
+                image_part_tags = [tag.strip() for tag in image_part_tags]
+                image_part_tags = Tag.objects.filter(name__in=image_part_tags)
+                image_part.tags.add(*image_part_tags)
+                added_parts += 1
 
-        logging.info(f'added_parts: {added_parts}')
+            logging.info(f'added_parts: {added_parts}')
 
-        logging.info(f'card_id: {card_id}; title: {title}; description: {description}; tags: {tags}; image: {card.image.url}')
+        # logging.info(f'card_id: {card_id}; title: {title}; description: {description}; tags: {tags}; image: {card.image.url}')
         return JsonResponse({"saved": True, "message": "Card saved successfully.", "card_id": card.id}, status=200)
